# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog post api",
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "ApiUser API"
                ],
                "summary": "Register",
                "operationId": "a6ee5da5a47c466565d321f664cc5d61",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "998902017747"
                                    },
                                    "fullname": {
                                        "type": "string",
                                        "example": "O'g'iloy"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check user **token** and added new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Qo'shildi"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "ApiUser API"
                ],
                "summary": "Login",
                "operationId": "f68a32310fa58f361d8560dd2c297508",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "998902017747"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check user **token** and added new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 11,
                                                "phone": "998902017747",
                                                "fullname": "O'g'iloy",
                                                "token": "htg387htv73gw3yg8v9g8yrfg923f4",
                                                "password": "$2y$12$lkQctPgpsD2bYXkgHPJLt.DLpAJwjS.uXlCpc8Bt39pGv4G8tcNTK",
                                                "created_at": "2023-08-29T04:39:24.000000Z",
                                                "updated_at": null
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Xush Kelibsiz!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get": {
            "get": {
                "tags": [
                    "ApiUser API"
                ],
                "summary": "GetAppUser",
                "operationId": "5951ff209222dc82a0c30340fa5e7e1a",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check user **token** and added new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 11,
                                                "phone": "998902017747",
                                                "fullname": "O'g'iloy",
                                                "token": "htg387htv73gw3yg8v9g8yrfg923f4",
                                                "password": "$2y$12$lkQctPgpsD2bYXkgHPJLt.DLpAJwjS.uXlCpc8Bt39pGv4G8tcNTK",
                                                "created_at": "2023-08-29T04:39:24.000000Z",
                                                "updated_at": null
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/list": {
            "get": {
                "tags": [
                    "ApiUser API"
                ],
                "summary": "GetUserlist",
                "operationId": "1496d8bbbba317c90dae81f9a84d8b51",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check user **token** and added new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": [
                                                {
                                                    "id": 11,
                                                    "phone": "998902017747",
                                                    "fullname": "O'g'iloy",
                                                    "token": "htg@htg387htv73gw3yg8v9g#%@8yr3f4",
                                                    "password": "$2y$12$lkQctPgpsD2bYXkgHPJLt.DLpAJwjS.uXlCpc8Bt39pGv4G8tcNTK",
                                                    "created_at": "2023-08-29T04:39:24.000000Z",
                                                    "updated_at": null
                                                },
                                                {
                                                    "id": 7,
                                                    "phone": "998902017747",
                                                    "fullname": "O'g'iloy",
                                                    "token": "htg387htv73^%@gw3yg#$%^28v9g8yrfg9",
                                                    "password": "$2y$12$lkQctPgpsD2bYXkgHPJLt.DLpAJwjS.uXlCpc8Bt39pGv4G8tcNTK",
                                                    "created_at": "2023-08-29T04:39:24.000000Z",
                                                    "updated_at": null
                                                }
                                            ]
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Here is your list"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}